Backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
flask-restx==1.3.0
mediapipe==0.10.14
opencv-python==4.10.0.84
numpy==1.26.4
Werkzeug==3.0.3
pytest==8.3.3
Pillow==10.4.0
replicate==0.25.1
requests==2.31.0
google-generativeai==0.3.2
PyJWT==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0

# ML Model Eğitimi için (opsiyonel)
tensorflow>=2.15.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
